+++ crates/cli/src/main.rs
@@ -22,7 +22,7 @@ use clap_complete::Shell;
 use wasefire_cli_tools::{action, fs};
 
 #[derive(Parser)]
-#[command(version, about)]
+#[command(name = "wasefire", version, about)]
 struct Flags {
     #[command(flatten)]
     options: Options,
@@ -44,14 +44,34 @@ enum Action {
     /// Lists the applets installed on a platform.
     AppletList,
 
-    /// Installs an applet on a platform.
-    AppletInstall,
+    #[group(id = "Action::AppletInstall")]
+    AppletInstall {
+        #[command(flatten)]
+        options: action::ConnectionOptions,
+        #[command(flatten)]
+        action: action::AppletInstall,
+        #[command(subcommand)]
+        command: Option<AppletInstallCommand>,
+    },
 
     /// Updates an applet on a platform.
     AppletUpdate,
 
-    /// Uninstalls an applet from a platform.
-    AppletUninstall,
+    #[group(id = "Action::AppletUninstall")]
+    AppletUninstall {
+        #[command(flatten)]
+        options: action::ConnectionOptions,
+        #[command(flatten)]
+        action: action::AppletUninstall,
+    },
+
+    #[group(id = "Action::AppletExitStatus")]
+    AppletExitStatus {
+        #[command(flatten)]
+        options: action::ConnectionOptions,
+        #[command(flatten)]
+        action: action::AppletExitStatus,
+    },
 
     // TODO(https://github.com/clap-rs/clap/issues/2621): We should be able to remove the explicit
     // group name.
@@ -84,6 +104,13 @@ enum Action {
         #[command(flatten)]
         action: action::PlatformReboot,
     },
+    #[group(id = "Action::PlatformLock")]
+    PlatformLock {
+        #[command(flatten)]
+        options: action::ConnectionOptions,
+        #[command(flatten)]
+        action: action::PlatformLock,
+    },
     #[group(id = "Action::PlatformRpc")]
     PlatformRpc {
         #[command(flatten)]
@@ -97,6 +124,16 @@ enum Action {
 
     /// Generates a shell completion file.
     Completion(Completion),
+}
+
+#[derive(clap::Subcommand)]
+enum AppletInstallCommand {
+    /// Waits until the applet exits.
+    #[group(id = "AppletInstallCommand::Wait")]
+    Wait {
+        #[command(flatten)]
+        action: action::AppletExitStatus,
+    },
 }
 
 #[derive(clap::Args)]
@@ -134,9 +171,24 @@ async fn main() -> Result<()> {
     let dir = std::env::current_dir()?;
     match flags.action {
         Action::AppletList => bail!("not implemented yet"),
-        Action::AppletInstall => bail!("not implemented yet"),
+        Action::AppletInstall { options, action, command } => {
+            let mut connection = options.connect().await?;
+            action.run(&mut connection).await?;
+            match command {
+                None => Ok(()),
+                Some(AppletInstallCommand::Wait { mut action }) => {
+                    action.wait.ensure_wait();
+                    action.run(&mut connection).await
+                }
+            }
+        }
         Action::AppletUpdate => bail!("not implemented yet"),
-        Action::AppletUninstall => bail!("not implemented yet"),
+        Action::AppletUninstall { options, action } => {
+            action.run(&mut options.connect().await?).await
+        }
+        Action::AppletExitStatus { options, action } => {
+            action.run(&mut options.connect().await?).await
+        }
         Action::AppletRpc { options, action } => action.run(&mut options.connect().await?).await,
         Action::PlatformList(x) => x.run().await,
         Action::PlatformUpdateMetadata { options } => {
@@ -149,6 +201,7 @@ async fn main() -> Result<()> {
         Action::PlatformReboot { options, action } => {
             action.run(&mut options.connect().await?).await
         }
+        Action::PlatformLock { options, action } => action.run(&mut options.connect().await?).await,
         Action::PlatformRpc { options, action } => action.run(&mut options.connect().await?).await,
         Action::RustAppletNew(x) => x.run().await,
         Action::RustAppletBuild(x) => x.run(dir).await,