+++ crates/runner-host/src/main.rs
@@ -21,7 +21,6 @@ use std::sync::Mutex;
 use anyhow::Result;
 use board::Board;
 use clap::Parser;
-use tokio::runtime::Handle;
 use tokio::select;
 use tokio::sync::mpsc::{channel, Receiver};
 use wasefire_board_api::Event;
@@ -97,6 +96,7 @@ async fn main() -> Result<()> {
     let storage = flags.flash_dir.join("storage.bin");
     let options = FileOptions { word_size: 4, page_size: 4096, num_pages: 16 };
     let storage = Some(FileStorage::new(&storage, options).unwrap());
+    board::applet::init(flags.flash_dir.join("applet.bin")).await;
     let (sender, receiver) = channel(10);
     *RECEIVER.lock().unwrap() = Some(receiver);
     #[cfg(feature = "web")]
@@ -169,11 +169,6 @@ async fn main() -> Result<()> {
         }
     });
     println!("Board initialized. Starting scheduler.");
-    #[cfg(feature = "wasm")]
-    const WASM: &[u8] = include_bytes!("../../../target/wasefire/applet.wasm");
-    #[cfg(feature = "wasm")]
-    Handle::current().spawn_blocking(|| Scheduler::<board::Board>::run(WASM)).await?;
-    #[cfg(feature = "native")]
-    Handle::current().spawn_blocking(|| Scheduler::<board::Board>::run()).await?;
-    Ok(())
+    // Not sure why Rust doesn't figure out this can't return (maybe async).
+    let _: ! = tokio::task::spawn_blocking(|| Scheduler::<board::Board>::run()).await?;
 }