+++ crates/board/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Major
 
+- Add `Api::Applet` as a required API for simple applet management
+- Remove `debug::Api::exit()`
 - Remove `api-platform{,-protocol,-update}` features making those APIs required
 - (Only when `api-platform-protocol` is used) Change `platform::protocol::Api::{enable,disable}()`
   to also control whether requests are accepted