+++ crates/protocol/CHANGELOG.md
@@ -8,6 +8,9 @@
 
 ### Minor
 
+- Add `PlatformLock` to lock a platform protocol
+- Add `AppletExitStatus` and `applet::ExitStatus` to get an applet exit status
+- Add `Applet{Install,Uninstall}` for applet management
 - Add `ConnectionExt::call_ref()` to share a request between calls
 - Add a `Service::NAME` constant with `host` feature
 - Add `PlatformUpdate{Metadata,Transfer}` calls and `transfer` module for platform updates